pyproject.toml
[tool.poetry]
name = "donkit-ragops"
version = "0.5.12"
description = "CLI agent for building RAG pipelines"
authors = ["Donkit AI <opensource@donkit.ai>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "donkit_ragops", from = "src" }]
include = [
    "src/donkit_ragops/compose/services/*.yml",
    "src/donkit_ragops/compose/templates/*",
    { path = "src/donkit_ragops/web/static/*", format = ["sdist", "wheel"] },
    { path = "src/donkit_ragops/web/static/**/*", format = ["sdist", "wheel"] }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
rich = "^13.7.1"
pydantic = "^2.7.4"
pydantic-settings = "^2.6.1"
sqlmodel = "^0.0.16"
python-dotenv = "^1.0.1"
httpx = "^0.28.1"
typer = "^0.17.4"
loguru = "^0.7.3"
openai = "^2.1.0"
google-genai = "^1.38.0"
langchain = "^0.3.27"
donkit-chunker = "0.1.9"
donkit-vectorstore-loader = "^0.1.9"
donkit-embeddings = "0.1.10"
langchain-openai = "^0.3.35"
langchain-community = "^0.3.31"
donkit-read-engine = "^0.2.7"
fastmcp = "^2.13.0.2"
donkit-ragops-api-gateway-client = "^0.2.1"
nest-asyncio = "^1.6.0"
keyring = "^25.6"
donkit-llm = "0.1.13"
prompt-toolkit = "^3.0.48"
toonify = {extras = ["pydantic"], version = "^1.5.1"}
tiktoken = "^0.12.0"
langdetect = "^1.0.9"
# Web interface dependencies
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.34"}
websockets = "^15.0"
python-multipart = "^0.0.20"
aiofiles = "^25.1"


[tool.poetry.scripts]
donkit-ragops = "donkit_ragops.cli:app"
donkit-ragops-mcp = "donkit_ragops.mcp.servers.donkit_ragops_mcp:main"
ragops-rag-planner = "donkit_ragops.mcp.servers.planner_server:main"
ragops-read-engine = "donkit_ragops.mcp.servers.read_engine_server:main"
ragops-chunker = "donkit_ragops.mcp.servers.chunker_server:main"
ragops-vectorstore-loader = "donkit_ragops.mcp.servers.vectorstore_loader_server:main"
ragops-checklist = "donkit_ragops.mcp.servers.checklist_server:main"
ragops-compose-manager = "donkit_ragops.mcp.servers.compose_manager_server:main"
ragops-rag-query = "donkit_ragops.mcp.servers.rag_query_server:main"
donkit-ragops-web = "donkit_ragops.web.app:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^6.0.0"
ruff = "^0.12.0"
pytest-asyncio = "^1.3.0"
# local dev deps
#donkit-ragops-api-gateway-client = { path = "../ragops-agent/shared/ragops-api-gateway-client", develop = true }
#donkit-llm = { path = "../ragops-agent/shared/donkit-llm", develop = true }
#donkit-chunker = { path = "../ragops-agent/shared/chunker", develop = true }
#donkit-vectorstore-loader = { path = "../ragops-agent/shared/vectorstore-loader", develop = true }
#donkit-read-engine = { path = "../ragops-agent/shared/read-engine", develop = true }
#donkit-embeddings = { path = "../ragops-agent/shared/embeddings", develop = true }


[tool.ruff]
line-length = 100
exclude = [".venv", "venv", "build", "dist", ".git", ".mypy_cache", ".ruff_cache", "**/prompts.py", "local_tools", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E402"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
strict = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-v --tb=short -p no:warnings --disable-warnings"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "asyncio: marks tests as async (using pytest-asyncio)",
]
